+++ src/bin/nile-rs/cli.rs
@@ -1,6 +1,6 @@
 use crate::commands::{
-    Call, Compile, CounterfactualAddress, Declare, DeclareV1, Deploy, GetAccounts, GetBalance,
-    GetNonce, Init, LegacyDeploy, Run, Send, Setup, Status,
+    Call, Clean, Compile, CounterfactualAddress, Declare, DeclareV1, Deploy, GetAccounts,
+    GetBalance, GetNonce, Init, LegacyDeploy, Run, Send, Setup, Status,
 };
 use clap::{Parser, Subcommand};
 
@@ -67,6 +67,9 @@ pub enum Commands {
 
     #[clap(about = "Query the registered accounts from the given network")]
     GetAccounts(GetAccounts),
+
+    #[clap(about = "Remove artifacts and contract deployments ")]
+    Clean(Clean),
 }
 
 #[test]